dcrm/mydb.py
import mysql.connector


dataBase = mysql.connector.connect(
    host = 'localhost',
    user = 'root',
    passwd = '3m1ly/2000',
    
    
)#create a database connection

#prepare a cursor object
cursorObject = dataBase.cursor()

#create a database

cursorObject.execute("CREATE DATABASE crmdatabase")

print("all done")