requirements.txt
Django==4.2.3
mysql_connector_repackaged==0.3.1
